add exceptionHandler/ fix services

+++ src/main/java/com/example/hop_oasis/service/data/BeerServiceImpl.java
@@ -1,10 +1,13 @@
 package com.example.hop_oasis.service.data;
 
+import com.example.hop_oasis.decoder.ImageCompressor;
 import com.example.hop_oasis.dto.BeerDto;
 import com.example.hop_oasis.dto.BeerInfoDto;
 import com.example.hop_oasis.dto.ImageDto;
 import com.example.hop_oasis.convertor.BeerInfoMapper;
 import com.example.hop_oasis.convertor.ImageMapper;
+import com.example.hop_oasis.hendler.exception.BeerNotFoundException;
+import com.example.hop_oasis.hendler.exception.ImageNotFoundException;
 import com.example.hop_oasis.model.Image;
 import com.example.hop_oasis.repository.BeerRepository;
 import com.example.hop_oasis.convertor.BeerMapper;
@@ -14,16 +17,14 @@ import com.example.hop_oasis.service.BeerService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
-
-import java.io.ByteArrayOutputStream;
 import java.io.IOException;
-import java.net.MalformedURLException;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Objects;
-import java.util.Optional;
-import java.util.zip.Deflater;
-import java.util.zip.Inflater;
+
+
+import static com.example.hop_oasis.hendler.exception.message.ExceptionMessage.*;
 
 @Service
 @RequiredArgsConstructor
@@ -33,9 +34,16 @@ public class BeerServiceImpl implements BeerService {
     private final BeerMapper beerMapper;
     private final BeerInfoMapper beerInfoMapper;
     private final ImageMapper imageMapper;
+    private final ImageCompressor imageCompressor;
+
     @Override
-    public void save(MultipartFile file, BeerDto beerDto) throws IOException {
-        byte[] bytesIm = compressImage(file.getBytes());
+    public void save(MultipartFile file, BeerDto beerDto)  {
+        byte[] bytesIm;
+        try {
+            bytesIm = imageCompressor.compressImage(file.getBytes());
+        } catch (IOException e) {
+            throw new ImageNotFoundException(IMAGE_COMPRESS_EXCEPTION,"");
+        }
         Image image = Image.builder()
                 .image(bytesIm)
                 .name(file.getOriginalFilename())
@@ -48,75 +56,19 @@ public class BeerServiceImpl implements BeerService {
         image.setBeer(beer);
         imageRepository.save(image);
     }
-    private byte[] compressImage(byte[] data) {
-        Deflater deflater = new Deflater();
-        deflater.setLevel(Deflater.BEST_COMPRESSION);
-        deflater.setInput(data);
-        deflater.finish();
-        ByteArrayOutputStream outputStream = new ByteArrayOutputStream(data.length);
-        byte[] tmp = new byte[2048];
-        try {
-            while (!deflater.finished()) {
-                int size = deflater.deflate(tmp);
-                outputStream.write(tmp, 0, size);
-            }
-            outputStream.close();
-        } catch (Exception e) {
-            System.out.println("Error beer problem");
-        }
-        return outputStream.toByteArray();
-    }
-    @Override
-    public void addImageToBeer(Long beerId, MultipartFile file) throws IOException {
-        byte[] image = compressImage(file.getBytes());
-
-        Image image1 = Image.builder()
-                .image(image)
-                .name(file.getOriginalFilename())
-                .build();
-
-        image1.setBeer(beerRepository.findById(beerId).orElse(null));
-        imageRepository.save(image1);
-    }
-    private static byte[] decompressImage(byte[] data) {
-        Inflater inflater = new Inflater();
-        inflater.setInput(data);
-        ByteArrayOutputStream outputStream = new ByteArrayOutputStream(data.length);
-        byte[] tmp = new byte[2048];
-        try {
-            while (!inflater.finished()) {
-                int count = inflater.inflate(tmp);
-                outputStream.write(tmp, 0, count);
-            }
-            outputStream.close();
-        } catch (Exception e) {
-            System.out.println("Error beer problem");
-        }
-        return outputStream.toByteArray();
-    }
     @Override
     public BeerInfoDto getBeerById(Long id) {
         Beer beer = beerRepository.findById(id).orElse(null);
         if (beer == null) {
-            throw new RuntimeException("Beer not found");
+            throw new BeerNotFoundException(BEER_NOT_FOUND, id);
         }
         return beerInfoMapper.toDto(beer);
     }
     @Override
-    public ImageDto getImageByName(String name) throws MalformedURLException {
-        Optional<Image> imageOp = imageRepository.findByName(name);
-        if (imageOp.isEmpty()) {
-            System.out.println("Image not found");
-        }
-        Image image = imageOp.get();
-        image.setImage(decompressImage(image.getImage()));
-        return imageMapper.toDto(image);
-    }
-    @Override
     public List<BeerInfoDto> getAllBeers() {
         List<Beer> beers = beerRepository.findAll();
-        if (beers == null) {
-            throw new RuntimeException("Beers not found");
+        if (beers.isEmpty()) {
+            throw new BeerNotFoundException(BEERS_NOT_FOUND, "");
         }
         return beerInfoMapper.toDtos(beers);
     }
@@ -124,7 +76,7 @@ public class BeerServiceImpl implements BeerService {
     public void update(BeerInfoDto beerInfo, Long id) {
         Beer beer = beerRepository.findById(id).orElse(null);
         if (beer == null) {
-            System.out.println("Beer not found");
+            throw new BeerNotFoundException(BEER_NOT_FOUND, id);
         }
         if (!beerInfo.getBeerName().isEmpty()) {
             beer.setBeerName(beerInfo.getBeerName());
@@ -151,6 +103,10 @@ public class BeerServiceImpl implements BeerService {
     }
     @Override
     public void delete(Long id) {
+        Beer beer = beerRepository.findById(id).orElse(null);
+        if (beer == null) {
+            throw new BeerNotFoundException(BEER_DELETED, id);
+        }
         beerRepository.deleteById(id);
     }
 }
